Move create_talk and create_talk_type to wafer.talks.tests.fixtures
