+++ wafer/talks/tests/test_views.py
@@ -8,28 +8,9 @@ from django.urls import reverse
 from wafer.tests.api_utils import SortedResultsClient
 from wafer.tests.utils import create_user
 from wafer.talks.models import (
-    Talk, TalkUrl, TalkType, ACCEPTED, REJECTED, SUBMITTED,
-    UNDER_CONSIDERATION, CANCELLED, PROVISIONAL)
-
-
-def create_talk_type(name):
-    """Create a talk type"""
-    return TalkType.objects.create(name=name)
-
-
-
-def create_talk(title, status, username, talk_type=None):
-    user = create_user(username)
-    talk = Talk.objects.create(
-        title=title, status=status, corresponding_author_id=user.id)
-    talk.authors.add(user)
-    talk.notes = "Some notes for talk %s" % title
-    talk.private_notes = "Some private notes for talk %s" % title
-    talk.save()
-    if talk_type:
-        talk.talk_type = talk_type
-        talk.save()
-    return talk
+    Talk, TalkUrl, ACCEPTED, REJECTED, SUBMITTED, UNDER_CONSIDERATION,
+    CANCELLED, PROVISIONAL)
+from wafer.talks.tests.fixtures import create_talk, create_talk_type
 
 
 def mock_avatar_url(self):
